Create server.py
Example flask app with two routes. The index route shows a simple form (when the /users route is commented out).

The second route has been added to demonstrate the pitfalls of attempting to rend a template on a POST request. Accessing the URL for the /users route with a POST method will return a 405 error.

An HTTP 405 Error indicates that a web browser has requested access to one of your pages, and your web server has recognized the request. However, the server has rejected the specific HTTP method it's using. As a result, your web browser can't access the requested web page.